+++ flutter_plugin_generator/lib/src/flutter_plugin_generator.dart
@@ -4,13 +4,19 @@ import 'package:flutter_plugin_annotations/flutter_plugin_annotations.dart';
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 
-/// This class reads the [MethodCallPlugin] annotation and the classes where it is applied.
-/// It will create a [MethodChannel] using the supplied [MethodCallPlugin.channelName] and
-/// a class with the prefix `_$`that implements the class where the annotation was used.
+/// This class reads the [FlutterPlugin] annotation and the classes where it is applied.
+/// If [MethodChannelFutures] is applied too, it will create a [MethodChannel] using the supplied
+/// [MethodChannelFutures.channelName] and a class with the prefix `_$`that extends the class
+/// where the annotation was used.
+///
 /// Example:
 ///
-/// ``` dart
-/// @MethodCallPlugin(channelName: "my channel name")
+/// ```dart
+///
+/// part 'platform_plugin.g.dart';
+///
+/// @FlutterPlugin()
+/// @MethodChannelFutures(channelName: "my channel name")
 /// abstract class PlatformPlugin {
 ///
 ///  Future<String> platform();
@@ -19,8 +25,8 @@ import 'package:source_gen/source_gen.dart';
 /// ```
 /// Will generate:
 ///
-///``` dart
-/// part of 'platform_plugin.dart';
+/// ```dart
+/// part of 'flutter_plugin.dart';
 ///
 /// **************************************************************************
 /// FlutterPluginGenerator
@@ -39,18 +45,65 @@ import 'package:source_gen/source_gen.dart';
 ///     return result;
 ///   }
 /// }
-///```
 ///
-/// /// See also:
+/// ```
+///
+/// If [EventChannelStream] annotation is applied to a field or getter of a class annotated with [FlutterPlugin]
+/// an static const [EventChannel] will be created.
+///
+/// Example:
+/// '''dart
+///
+///  part 'platform_plugin.g.dart';
+///
+///  @FlutterPlugin()
+///  abstract class PlatformPlugin {
+///   Stream<String> get platform;
+///
+///   static PlatformPlugin create() {
+///     return _$PlatformPlugin();
+///   }
+/// }
+///
+/// ```
+///
+/// ```dart
+/// part of 'platform_plugin.dart';
+///
+/// **************************************************************************
+/// FlutterPluginGenerator
+/// **************************************************************************
+///
+/// class _$PlatformPlugin extends PlatformPlugin {
+///  _$PlatformPlugin();
+///
+///  static const EventChannel _platformEventChannel =
+///      const EventChannel('my event channel');
+///
+///  final Stream<dynamic> _platform =
+///      _platformEventChannel.receiveBroadcastStream();
+///
+///  @override
+///  Stream<String> get platform {
+///    return _platform.map((result) {
+///      return result;
+///    });
+///  }
+/// }
+///
+/// ```
+///
+/// See also:
 ///
 ///  * [https://pub.dev/packages/flutter_plugin_annotations], counter part library to this one.
-class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
+///
+class FlutterPluginGenerator extends GeneratorForAnnotation<FlutterPlugin> {
   const FlutterPluginGenerator();
 
   @override
   generateForAnnotatedElement(
       Element element, ConstantReader annotation, BuildStep buildStep) {
-    final template = generatePluginTemplate(element, annotation);
+    final template = generatePluginTemplate(element as ClassElement);
     return template;
   }
 
@@ -77,78 +130,85 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
   }
 
   String declareMethods(
-    ClassElement element,
+    ClassElement rootElement,
     List<SupportedPlatform> pluginPlatforms,
   ) {
     final buffer = StringBuffer();
 
-    final methods = element.methods.where((method) {
+    final futureMethods = rootElement.methods.where((method) {
       return method.isAbstract &&
           method.isPublic &&
           method.returnType.isDartAsyncFuture;
     }).toList();
 
-    final fields = element.fields
+    final futureFields = rootElement.fields
         .where((field) => field.isPublic && field.type.isDartAsyncFuture)
         .toList();
 
-    fields.forEach((field) {
-      buffer.write(declareGetter(field, pluginPlatforms));
+    futureFields.forEach((field) {
+      buffer.write(declareFutureGetter(field, pluginPlatforms));
     });
 
-    methods.forEach((method) {
-      buffer.write(declareMethod(method, pluginPlatforms));
+    futureMethods.forEach((method) {
+      buffer.write(declareFutureMethod(method, pluginPlatforms));
     });
 
     return buffer.toString();
   }
 
-  String declareGetter(
+  String declareFutureGetter(
     FieldElement field,
     List<SupportedPlatform> pluginPlatforms,
   ) {
-    final buffer = StringBuffer();
     final fieldPlatforms =
         findSupportedPlatforms(field.metadata + field.getter.metadata);
 
     final fieldName = field.displayName;
     final fieldType = field.type.displayName;
 
-    buffer.writeln('@override');
-    buffer.writeln('$fieldType get $fieldName async {');
-
-    buffer.writeln(
-        processSupportedPlatforms(fieldName, pluginPlatforms, fieldPlatforms));
+    final platformsOnly = processSupportedPlatforms(
+      fieldName,
+      pluginPlatforms: pluginPlatforms,
+      elementPlatforms: fieldPlatforms,
+    );
 
     final innerType = (field.type as ParameterizedType).typeArguments[0];
     final invokeMethod = selectInvokeMethod(innerType);
 
     final resultMapped = mapResultToDart(innerType, false);
 
-    buffer.writeln(
-        'final result = await _methodChannel.$invokeMethod(\'$fieldName\');');
-    buffer.writeln('$resultMapped');
-    buffer.writeln('}');
-
-    return buffer.toString();
+    return '''
+    
+          @override
+          $fieldType get $fieldName async {
+            
+              $platformsOnly
+              
+              final result = await _methodChannel.$invokeMethod(\'$fieldName\');
+              
+              return $resultMapped;
+              
+           }
+    
+    ''';
   }
 
-  String declareMethod(
+  String declareFutureMethod(
     MethodElement method,
     List<SupportedPlatform> pluginPlatforms,
   ) {
-    final buffer = StringBuffer();
     final methodPlatforms = findSupportedPlatforms(method.metadata);
 
     final methodName = method.displayName;
 
     final methodParams = declareParams(method.parameters);
     final methodReturnType = method.returnType.displayName;
-    buffer.writeln('@override');
-    buffer.writeln('$methodReturnType $methodName($methodParams) async {');
 
-    buffer.writeln(processSupportedPlatforms(
-        methodName, pluginPlatforms, methodPlatforms));
+    final platformsOnly = processSupportedPlatforms(
+      methodName,
+      pluginPlatforms: pluginPlatforms,
+      elementPlatforms: methodPlatforms,
+    );
 
     final innerType = (method.returnType as ParameterizedType).typeArguments[0];
     final invokeMethod = selectInvokeMethod(innerType);
@@ -157,12 +217,17 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
 
     final resultMapped = mapResultToDart(innerType, false);
 
-    buffer.writeln(
-        'final result = await _methodChannel.$invokeMethod(\'$methodName\'$separator $invokeParams);');
-    buffer.writeln('$resultMapped');
-    buffer.writeln('}');
+    return '''
+            @override
+            $methodReturnType $methodName($methodParams) async {
 
-    return buffer.toString();
+                $platformsOnly
+                
+                final result = await _methodChannel.$invokeMethod(\'$methodName\'$separator $invokeParams);
+                
+                return $resultMapped;
+            }
+    ''';
   }
 
   String declareParams(List<ParameterElement> parameters) {
@@ -175,8 +240,7 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
       if (firstNamed == param) {
         buffer.writeln('{');
       }
-      buffer.write(param.type.displayName + ' ' + param.displayName);
-      buffer.write(', ');
+      buffer.write(' ${param.type.displayName}  ${param.displayName}, ');
     });
     if (firstNamed != null) {
       buffer.writeln('}');
@@ -209,7 +273,6 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
     return 'invokeMapMethod<String, dynamic>';
   }
 
-
   bool isCoreDartType(DartType type) {
     return type.isDartCoreString ||
         type.isDartCoreBool ||
@@ -225,17 +288,20 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
   /// TODO : make this type mapping recursive to add support to things like List<List<MyData>> and Map<List<MyData>, List<MyOtherData>>
   String mapResultToDart(DartType type, bool includeExtraCasting) {
     if (isCoreDartType(type)) {
-      return 'return result;';
+      return 'result';
     } else if (type.isDartCoreList) {
       final innerType = (type as ParameterizedType).typeArguments[0];
       final extraCasting =
           includeExtraCasting ? 'List.castFrom(result)' : 'result';
       if (isCoreDartType(innerType)) {
-        return 'return $extraCasting;';
+        return '$extraCasting';
       } else {
-        return 'return $extraCasting' +
-            '.map((item) => Map<String, dynamic>.from(item))' +
-            '.map((item) => ${innerType.displayName}.fromJson(item)).toList();';
+        return '''
+              $extraCasting
+                          .map((item) => Map<String, dynamic>.from(item))
+                          .map((item) => ${innerType.displayName}.fromJson(item)).toList()
+        
+        ''';
       }
     }
     if (type.isDartCoreMap) {
@@ -243,7 +309,7 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
       final valueType = (type as ParameterizedType).typeArguments[1];
 
       if (isCoreDartType(keyType) && isCoreDartType(valueType)) {
-        return 'return Map<${keyType.displayName}, ${valueType.displayName}>.from(result);';
+        return 'Map<${keyType.displayName}, ${valueType.displayName}>.from(result)';
       } else {
         final key =
             isCoreDartType(keyType) ? 'key' : 'Map<String, dynamic>.from(key)';
@@ -262,24 +328,29 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
             isCoreDartType(valueType) ? valueType.displayName : 'dynamic';
         final extraCasting =
             includeExtraCasting ? 'Map<$kk, $vk>.from(result)' : 'result';
-        return ''' return $extraCasting
+        return '''
+                  $extraCasting
                             .map((key, value) => MapEntry(
-                              $key,
-                              $value,
-                            ))
+                                $key,
+                                $value,
+                              ),
+                            )
                             .map((key, value) => MapEntry(
-                              $key2, 
-                              $value2,
-                            ));''';
+                                $key2, 
+                                $value2,
+                              ),
+                            )
+                            
+             ''';
       }
     } else {
       if (includeExtraCasting) {
         return '''
         final item = Map<String, dynamic>.from(result);
-        return ${type.displayName}.fromJson(item);''';
+        ${type.displayName}.fromJson(item)''';
       } else {
         return '''
-        return ${type.displayName}.fromJson(result);''';
+        ${type.displayName}.fromJson(result)''';
       }
     }
   }
@@ -310,7 +381,7 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
         } else {
           final key = isCoreDartType(keyType) ? 'key' : 'key.toJson()';
           final value = isCoreDartType(valueType) ? 'value' : 'value.toJson()';
-          return '${param.displayName}.map((key, value) => MapEntry($key, $value))';
+          return '${param.displayName}.map((key, value) => MapEntry($key, $value,),)';
         }
       } else {
         return '${param.displayName}.toJson()';
@@ -339,7 +410,7 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
             final key = isCoreDartType(keyType) ? 'key' : 'key.toJson()';
             final value =
                 isCoreDartType(valueType) ? 'value' : 'value.toJson()';
-            return '\'${param.displayName}\' : ${param.displayName}.map((key, value) => MapEntry($key, $value))';
+            return '\'${param.displayName}\' : ${param.displayName}.map((key, value) => MapEntry($key, $value,),)';
           }
         } else {
           return '\'${param.displayName}\' : ${param.displayName}.toJson()';
@@ -350,11 +421,12 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
     }
   }
 
-  String findStreamChannelName(List<ElementAnnotation> metadata) {
+  String findChannelName(
+      List<ElementAnnotation> metadata, String annotationName) {
     final annotation = metadata.firstWhere(
         (annotation) =>
             annotation.computeConstantValue().type.displayName ==
-            'EventChannelStream',
+            annotationName,
         orElse: () => null);
 
     return annotation == null
@@ -366,18 +438,20 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
     return type.displayName.startsWith('Stream<');
   }
 
-  String generatePluginTemplate(Element element, ConstantReader annotation) {
-    final rootElement = (element as ClassElement);
-    final channelName = annotation.read('channelName').stringValue;
-    final className = element.displayName;
+  String generatePluginTemplate(ClassElement rootElement) {
+    final channelName =
+        findChannelName(rootElement.metadata, 'MethodChannelFutures');
+    final className = rootElement.displayName;
 
-    final pluginPlatforms = findSupportedPlatforms(element.metadata);
+    final factory = createFactory(rootElement,
+        className: className, methodChannelName: channelName);
 
-    final factory = createFactory(rootElement, className, channelName);
+    final pluginPlatforms = findSupportedPlatforms(rootElement.metadata);
 
     final streamGetters = createStreamGetters(rootElement, pluginPlatforms);
 
-    final methods = declareMethods(rootElement, pluginPlatforms);
+    final methods =
+        channelName != null ? declareMethods(rootElement, pluginPlatforms) : '';
 
     return '''
     
@@ -393,11 +467,17 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
     ''';
   }
 
-  String createFactory(
-      ClassElement element, String className, String channelName) {
+  String createFactory(ClassElement rootElement,
+      {String className, String methodChannelName}) {
+    if (methodChannelName == null) {
+      return '''
+        _\$$className();
+      ''';
+    }
+
     final pathRegexp = RegExp('(\{.*?\})');
     final groups =
-        pathRegexp.allMatches(channelName).map((match) => match.group(0));
+        pathRegexp.allMatches(methodChannelName).map((match) => match.group(0));
 
     final constructorParameters = groups
         .map((group) => group.replaceAll('{', '').replaceAll('}', ''))
@@ -413,33 +493,34 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
     final factory = oneChannelByInstance
         ? '''
       
-       final MethodChannel _methodChannel;
+        final MethodChannel _methodChannel;
        
-      factory _\$$className({$constructorParameters}) {
+        factory _\$$className({$constructorParameters}) {
        
-          final channelName = \'$channelName\'.$replacements;
+            final channelName = \'$methodChannelName\'.$replacements;
           
-          return _\$$className.private(MethodChannel(channelName));
-      }
-      _\$$className.private(this._methodChannel);
+            return _\$$className.private(MethodChannel(channelName));
+        }
+        
+        _\$$className.private(this._methodChannel);
     '''
         : '''
         
-        static const MethodChannel _methodChannel = const MethodChannel('$channelName');
+         static const MethodChannel _methodChannel = const MethodChannel('$methodChannelName');
         
-        _\$$className();
+         _\$$className();
+        
       ''';
     return factory;
   }
 
-  String processSupportedPlatforms(
-      String elementName,
-      List<SupportedPlatform> pluginSupportedPlatforms,
-      List<SupportedPlatform> methodSupportedPlatforms) {
+  String processSupportedPlatforms(String elementName,
+      {List<SupportedPlatform> pluginPlatforms,
+      List<SupportedPlatform> elementPlatforms}) {
     final buffer = StringBuffer();
 
-    pluginSupportedPlatforms.forEach((platform) {
-      if (!methodSupportedPlatforms.contains(platform)) {
+    pluginPlatforms.forEach((platform) {
+      if (!elementPlatforms.contains(platform)) {
         final name = platformName(platform);
         buffer.writeln('''
         if (Platform.is$name)
@@ -451,37 +532,46 @@ class FlutterPluginGenerator extends GeneratorForAnnotation<MethodCallPlugin> {
   }
 
   String createStreamGetters(
-    ClassElement element,
+    ClassElement rootElement,
     List<SupportedPlatform> pluginPlatforms,
   ) {
-    return element.fields
+    return rootElement.fields
         .where((field) => isDartStream(field.type))
         .map((field) {
-      final channelName =
-          findStreamChannelName(field.metadata + field.getter.metadata);
+      final eventChannelName = findChannelName(
+          field.metadata + field.getter.metadata, 'EventChannelStream');
 
       final fieldPlatforms =
           findSupportedPlatforms(field.metadata + field.getter.metadata);
+
       final platformOnly = processSupportedPlatforms(
-          field.displayName, pluginPlatforms, fieldPlatforms);
+        field.displayName,
+        pluginPlatforms: pluginPlatforms,
+        elementPlatforms: fieldPlatforms,
+      );
+
       final innerType = (field.type as ParameterizedType).typeArguments[0];
+
       final result = mapResultToDart(innerType, true);
 
-      if (channelName != null && channelName.isNotEmpty) {
+      if (eventChannelName != null) {
         return '''
-        static const EventChannel _${field.displayName}EventChannel = const EventChannel('$channelName');
         
-        final _${field.displayName} = _${field.displayName}EventChannel.receiveBroadcastStream();
+        static const EventChannel _${field.displayName}EventChannel = const EventChannel('$eventChannelName');
+        
+        final Stream<dynamic> _${field.displayName} = _${field.displayName}EventChannel.receiveBroadcastStream();
         
         @override
         ${field.type.displayName} get ${field.displayName} {
+          
             $platformOnly
-        return  _${field.displayName}.map((result){  $result });
+          
+          return  _${field.displayName}.map((result){  return $result; });
         }
+        
       ''';
       } else {
-        throw UnsupportedError(
-            'No @EventChannelStream annotation found for ${field.displayName}');
+        return '';
       }
     }).join('\n');
   }